Replaced CGIHandler class with ResponseFromCgi class and it returns from HttpRequestChecker class, but it needs some parameters like servername, port and also the path must be generated by default
