+++ Makefile
@@ -14,6 +14,7 @@ SRCS		=	src/main.cpp src/Server/create_http.cpp \
 				src/ReturnStatus/Response/NoContent204.cpp \
 				src/ReturnStatus/Response/InternalServerError500.cpp \
 				src/ReturnStatus/Response/Forbidden403.cpp \
+				src/ReturnStatus/Response/ResponseFromCgi.cpp \
 				src/ReturnStatus/Response/OK200.cpp \
 				src/ReturnStatus/Response/CommonFuncs.cpp \
 				src/CGI/cgi.cpp src/CGIHandler/CGIHandler.cpp \
