+++ inc/webserv.hpp
@@ -45,6 +45,8 @@
 #include "../src/ReturnStatus/Response/NoContent204.hpp"
 #include "../src/ReturnStatus/Response/InternalServerError500.hpp"
 #include "../src/ReturnStatus/Response/OK200.hpp"
+#include "../src/ReturnStatus/Response/ResponseFromCgi.hpp"
+
 
 // 
 #include "../src/CGIHandler/CGIHandler.hpp"
@@ -61,7 +63,7 @@
 #define GREEN_TEXT "\033[1;32m"
 // 
 #define CGI_BIN "cgi-bin/php-cgi"
-#define ROOT_FOLDER "/Users/aessaoud/Desktop/webserv/var/www/html"
+#define ROOT_FOLDER "/Users/aessaoud/Desktop/projects/webserv/var/www/html"
 #define PATH "/Users/aessaoud/goinfre/homebrew/bin:/Users/aessaoud/goinfre/homebrew/bin:/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin:/usr/local/munki"
 
 class ReturnStatus;
