+++ src/CGIHandler/CGIHandler.hpp
@@ -12,19 +12,19 @@ typedef struct
 class CGIHandler
 {
 
-public:
-	CGIHandler();
-	CGIHandler(RequestData request);
-	CGIHandler(CGIHandler const &src);
-	~CGIHandler();
-	CGIHandler &operator=(CGIHandler const &rhs);
-	// 
-	std::string process();
-	void init_env(RequestData request);
-	// 
-private:
-	std::map<std::string, std::string> keyValue;
-	std::vector<const char *> env_vec;
+// public:
+// 	CGIHandler();
+// 	CGIHandler(RequestData request);
+// 	CGIHandler(CGIHandler const &src);
+// 	~CGIHandler();
+// 	CGIHandler &operator=(CGIHandler const &rhs);
+// 	// 
+// 	std::string process();
+// 	void init_env(RequestData request);
+// 	// 
+// private:
+// 	std::map<std::string, std::string> keyValue;
+// 	std::vector<const char *> env_vec;
 };
 
 #endif /* ****************************************************** CGIHANDLER_H */
